+++ src/Products/PloneMeeting/browser/async.py
@@ -76,7 +76,7 @@ class Discuss(BrowserView):
         if discussAction == 'ask':
             # I must send a mail to MeetingManagers for notifying them that a reviewer
             # wants to discuss this item.
-            sendMailEnabled = item.sendMailIfRelevant('askDiscussItem', 'MeetingManager', isRole=True)
+            sendMailEnabled = item.sendMailIfRelevant('askDiscussItem', 'meetingmanagers', isSuffix=True)
             if sendMailEnabled:
                 msgId = 'to_discuss_ask_mail_sent'
             else: